rename FCL.py to FL.py
I am aware that this potentially breaks backwards compatibility for certain people but there is nothing cluster in that file. Adding clusters requires a cluster definition so I'd suggest adding FCLs under chimera.py and creating a pegasus.py if one wants chimera shaped clusters
