+++ dimod/generators/fl.py
